+++ packages/store/tests/issues/issue-2180-dispatching-in-effect.spec.ts
@@ -3,7 +3,8 @@ import {
   Injectable,
   effect,
   provideExperimentalZonelessChangeDetection,
-  signal
+  signal,
+  untracked
 } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
 import { Action, NgxsModule, Selector, State, StateContext, Store } from '@ngxs/store';
@@ -48,7 +49,7 @@ describe('State per signal', () => {
     constructor(private store: Store) {
       effect(() => {
         const value = this.value();
-        store.dispatch(new SetNumber(value));
+        untracked(() => store.dispatch(new SetNumber(value)));
       });
     }
